pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "document-intelligence-toolkit"
version = "0.1.0"
description = "Enterprise-grade document ingestion, extraction, summarization, and comparison toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # LangChain core + integrations
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    # Document parsing
    "pdfplumber>=0.11.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    # Token management
    "tiktoken>=0.7.0",
    # Diff / comparison utilities
    "deepdiff>=7.0.0",
    # UI
    "streamlit>=1.37.0",
    # Config & environment
    "python-dotenv>=1.0.0",
    # Data handling
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "ui*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"
